refactor(backend): extract VCL admin and tracking routes into separate modules
